+++ src/PerpetualIntelligence.Terminal/Commands/Handlers/IExceptionHandler.cs
@@ -5,15 +5,20 @@
     https://terms.perpetualintelligence.com/articles/intro.html
 */
 
-using PerpetualIntelligence.Shared.Abstractions;
 using System;
+using System.Threading.Tasks;
 
 namespace PerpetualIntelligence.Terminal.Commands.Handlers
 {
     /// <summary>
     /// An abstraction to handle <see cref="Exception"/>.
     /// </summary>
-    public interface IExceptionHandler : IHandlerNoResult<ExceptionHandlerContext>
+    public interface IExceptionHandler
     {
+        /// <summary>
+        /// Handles <see cref="Exception"/> asynchronously.
+        /// </summary>
+        /// <param name="context">The exception context.</param>
+        public Task HandleAsync(ExceptionHandlerContext context);
     }
 }