+++ src/PerpetualIntelligence.Terminal/Commands/Providers/IDefaultOptionProvider.cs
@@ -5,14 +5,20 @@
     https://terms.perpetualintelligence.com/articles/intro.html
 */
 
-using PerpetualIntelligence.Shared.Abstractions;
+using System.Threading.Tasks;
 
 namespace PerpetualIntelligence.Terminal.Commands.Providers
 {
     /// <summary>
-    /// An abstraction to provide an option's default value.
+    /// An abstraction to provide a default option.
     /// </summary>
-    public interface IDefaultOptionProvider : IProvider<DefaultOptionProviderContext, DefaultOptionProviderResult>
+    public interface IDefaultOptionProvider
     {
+        /// <summary>
+        /// Provides a default option asynchronously.
+        /// </summary>
+        /// <param name="defaultOptionValueProviderContext">The default option context.</param>
+        /// <returns>The <see cref="DefaultOptionProviderResult"/> instance.</returns>
+        Task<DefaultOptionProviderResult> ProvideAsync(DefaultOptionProviderContext defaultOptionValueProviderContext);
     }
 }