+++ test/PerpetualIntelligence.Terminal.Test/Licensing/LicenseExtractorOfflineTests.cs
@@ -294,7 +294,7 @@ namespace PerpetualIntelligence.Terminal.Licensing
         public async Task ExtractFromJsonAsync_ValidKey_ShouldContainClaimsAsync()
         {
             // Before extract get should be null
-            License? licenseFromGet = await licenseExtractor.GetLicenseAsync();
+            License? licenseFromGet = await licenseExtractor.GetAsync();
             licenseFromGet.Should().BeNull();
 
             terminalOptions.Licensing.LicenseKey = testLicPath;
@@ -358,7 +358,7 @@ namespace PerpetualIntelligence.Terminal.Licensing
             result.License.Price.Yearly.Should().Be(35629.0);
 
             // After extract and Get should return the correct license
-            licenseFromGet = await licenseExtractor.GetLicenseAsync();
+            licenseFromGet = await licenseExtractor.GetAsync();
             licenseFromGet.Should().NotBeNull();
             licenseFromGet.Should().BeSameAs(result.License);
         }